betty errors fixed,5and 8 to bestill fixed

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -7,10 +7,11 @@
  *
  * Return: number of (nodes)elements in the list
  */
+
 size_t print_listint(const listint_t *h)
+
 {
 	size_t i = 0;
-        	
 
 	while (h != NULL)
 	{
@@ -18,5 +19,5 @@ size_t print_listint(const listint_t *h)
 		h = h->next;
 		i++;
 	}
-	return (int)i;
+	return (i);
 }
